Refactor dataset, transform etc
