+++ road_roughness_prediction/dataset/__init__.py
@@ -1,0 +1,3 @@
+'''Dataset module'''
+from .surface_category import SurfaceCategoryDataset
+from .surface_category import SurfaceCategoryDatasetFactory
