+++ tests/test_transformations.py
@@ -4,26 +4,11 @@ from pathlib import Path
 import pytest
 from PIL import Image
 
-from albumentations import Blur
-from albumentations import Compose
-from albumentations import IAAAdditiveGaussianNoise
-from albumentations import RandomCrop
-
 import numpy as np
 import matplotlib.pyplot as plt
 
-from road_roughness_prediction.tools.transformations import Rescale
-
-
-OUTPUT_SIZE = 256
-
-
-TRANSFORM = Compose([
-    IAAAdditiveGaussianNoise(scale=(0.01 * 255, 0.15 * 255.)),
-    Blur(blur_limit=4),
-    Rescale(OUTPUT_SIZE),
-    RandomCrop(OUTPUT_SIZE, OUTPUT_SIZE),
-])
+from road_roughness_prediction.config import Config
+from road_roughness_prediction.dataset.transformations import TransformFactory
 
 
 def _get_images(paths):
@@ -31,11 +16,12 @@ def _get_images(paths):
 
 
 class TestTransformations:
-
+    config = Config()
     root = Path('tests/resources/surfaces/')
     paths = list(root.glob('**/*.jpg'))
     images = _get_images(paths)
-    transforms = [TRANSFORM]
+    transforms = [TransformFactory(config)]
+    image_transforms = [tr.image_transform for tr in transforms]
 
     def _plot_transformation(self, transform, images, paths):
         for image, path in zip(images, paths):
@@ -54,10 +40,10 @@ class TestTransformations:
         '''Test image given from command line'''
         paths = [Path(image_path)]
         images = _get_images(paths)
-        for transform in self.transforms:
+        for transform in self.image_transforms:
             self._plot_transformation(transform, images, paths)
 
     @pytest.mark.interactive
     def test_transformations(self):
-        for transform in self.transforms:
+        for transform in self.image_transforms:
             self._plot_transformation(transform, self.images, self.paths)
